refactor: seperated client-side components from pages; improved image compression logic
- Fixed image paths and asset import
- Added comments to divide file into subgroups for clarity
- Added metadata for improved SEO
