+++ src/app/(static)/about/page.tsx
@@ -1,64 +1,92 @@
-"use client";
-
 import {
+  ISATechDecorationCenter,
   ISATechDecorationLeft,
   ISATechDecorationRight,
 } from "@/components/assets/decorations";
+import { ISATechLogoMark } from "@/components/assets/logos";
 import PerlinNoiseTexture from "@/components/shaders/perlin";
-import { BlobsAnimatedBackground, BlobsConfig } from "@/components/ui/blobs";
-import { Button } from "@/components/ui/button";
-import {
-  Carousel,
-  CarouselContent,
-  CarouselItem,
-  CarouselApi,
-} from "@/components/ui/carousel";
-import {
-  ArrowLeftIcon,
-  ArrowRightIcon,
-  RocketIcon,
-  TargetIcon,
-} from "lucide-react";
+import { RocketIcon, TargetIcon } from "lucide-react";
+import { Metadata } from "next";
 import Image from "next/image";
-import { useEffect, useState } from "react";
+import AboutUsAdvisersSection, { AdviserProps } from "./carousel";
 
-const advisers = [
+/**
+ * ################################################################################
+ * ################################### METADATA ###################################
+ * ################################################################################
+ */
+export const metadata: Metadata = {
+  title: "About Us | ISATech Society",
+  description:
+    "Learn more about ISATech Society, our mission, vision, and the team behind the innovation.",
+  openGraph: {
+    title: "About Us | ISATech Society",
+    description:
+      "Learn more about ISATech Society, our mission, vision, and the team behind the innovation.",
+    url: "https://isatech.club/about",
+    siteName: "ISATech Society",
+    type: "website",
+  },
+};
+
+/**
+ * ################################################################################
+ * #################################### CONFIG ####################################
+ * ################################################################################
+ */
+const advisers: AdviserProps[] = [
   {
     name: "Carmello V. Ambut, Ed. D.",
     title: "VP for Research and Extension",
-    image: "/assets/advisers/adviser-ambut.png",
+    image: "/assets/advisers/ambut.png",
+    imageWidth: 903,
+    imageHeight: 1430,
+    imageSize:
+      "(min-width: 1640px) calc(1.1vw + 281px), (min-width: 1280px) calc(3.53vw + 258px), (min-width: 1040px) calc(40.45vw - 56px), (min-width: 980px) 47.5vw, (min-width: 780px) calc(55vw - 31px), (min-width: 680px) 82.5vw, calc(111.39vw - 55px)",
   },
   {
     name: "Naci John C. Trance",
     title: "Director, Intellectual Property Management Office",
-    image: "/assets/advisers/adviser-trance.png",
+    image: "/assets/advisers/trance.png",
+    imageWidth: 1116,
+    imageHeight: 1430,
+    imageSize:
+      "(min-width: 1280px) 399px, (min-width: 1040px) calc(68.18vw - 265px), (min-width: 980px) 62.5vw, (min-width: 780px) calc(72.78vw - 50px), (min-width: 680px) 106.25vw, calc(143.06vw - 57px)",
+  },
+  {
+    name: "Karlo S. Sira",
+    title: "Chair, Creative Works Section - IPMO",
+    image: "/assets/advisers/sira.png",
+    imageWidth: 1116,
+    imageHeight: 1430,
+    imageSize:
+      "(min-width: 1280px) 370px, (min-width: 1040px) calc(63.18vw - 244px), (min-width: 980px) 57.5vw, (min-width: 780px) calc(67.78vw - 47px), (min-width: 680px) 101.25vw, calc(133.06vw - 53px)",
   },
   {
     name: "Rayjand T. Gellamucho",
-    title: "Adviser, ISATech Society",
-    image: "/assets/advisers/adviser-gellamucho.png",
+    title: "General Manager, KWADRA TBI",
+    image: "/assets/advisers/gellamucho.png",
+    imageWidth: 1116,
+    imageHeight: 1430,
+    imageSize:
+      "(min-width: 1280px) 370px, (min-width: 1040px) calc(63.18vw - 244px), (min-width: 980px) 57.5vw, (min-width: 780px) calc(67.78vw - 47px), (min-width: 680px) 101.25vw, calc(133.06vw - 53px)",
   },
   {
     name: "John Joseph L. Tabladillo",
-    title: "Co-Adviser, ISATech Society",
-    image: "/assets/advisers/adviser-tabladillo.png",
+    title: "Project Technical Assistant I, UMWAD Western Visayas",
+    image: "/assets/advisers/tabladillo.png",
+    imageWidth: 813,
+    imageHeight: 1430,
+    imageSize:
+      "(min-width: 1660px) calc(1.04vw + 252px), (min-width: 1280px) 279px, (min-width: 1040px) calc(36.36vw - 49px), (min-width: 980px) 42.5vw, (min-width: 780px) calc(50vw - 32px), (min-width: 680px) 75vw, calc(100vw - 48px)",
   },
 ];
 
-const adviserSectionBG: BlobsConfig[] = [
-  {
-    id: "default-blob-2",
-    top: "-10rem",
-    left: "-10rem",
-    animateX: [0, -30, 0],
-    animateY: [0, -40, 0],
-    duration: 6,
-    colorClass: "bg-secondary/60",
-    sizeClass: "h-96 w-96",
-    blurClass: "blur-[100px]",
-  },
-];
-
+/**
+ * ################################################################################
+ * ################################## COMPONENTS ##################################
+ * ################################################################################
+ */
 function AboutUsHeroSection() {
   return (
     <section
@@ -79,20 +107,18 @@ function AboutUsHeroSection() {
         {/* 4H Decoration */}
         <div className="hidden w-fit items-center justify-center sm:flex sm:w-full">
           <Image
-            src="/assets/isatech-decoration-4h-landscape.png"
-            alt="Logo"
-            width={280}
-            height={330}
+            src="/assets/decorations/4h-vertical.png"
+            alt="4H Vertical Pose"
+            width={279}
+            height={329}
             className="h-32 w-fit md:h-64"
+            sizes="(min-width: 780px) 217px, 109px"
           />
         </div>
         {/* ISATech Logo Decoration */}
-        <Image
-          src="/assets/isatech-decoration-2.svg"
-          alt="Logo"
-          width={200}
-          height={305}
+        <ISATechDecorationCenter
           className="absolute top-0 right-0 -z-1 h-full w-auto"
+          color="#203C90"
         />
       </div>
     </section>
@@ -112,21 +138,17 @@ function AboutUsDescriptionSection() {
       </div>
       <div className="flex w-full max-w-6xl flex-col items-center">
         {/* Text Content */}
-        <div className="bg-accent relative flex w-full max-w-6xl flex-col items-center justify-between gap-4 rounded-3xl px-6 lg:px-20">
+        <div className="bg-accent relative flex w-full max-w-6xl flex-col items-center justify-between gap-4 overflow-hidden rounded-3xl px-6 lg:px-20">
           <Image
-            src="/assets/isatech-group-photo-1.jpg"
-            alt="About Us"
+            src="/assets/decorations/tradeanovate-grouphoto.jpg"
+            alt="Tradeanovate Group Photo"
             width={1695}
             height={706}
-            className="absolute h-full w-auto rounded-3xl object-cover opacity-30"
+            sizes="(min-width: 1360px) 1153px, (min-width: 1040px) calc(80.33vw + 78px), (min-width: 960px) 709px, (min-width: 740px) 751px, (min-width: 600px) calc(-47.5vw + 1091px), (min-width: 400px) calc(-63.89vw + 1179px), calc(-298.75vw + 2054px)"
+            className="absolute h-full w-auto rounded-3xl object-cover opacity-30 lg:h-auto lg:w-full"
           />
           <div className="flex flex-col items-center justify-center gap-4 py-16 lg:py-32">
-            <Image
-              src="/assets/isatech-icon-dark.svg"
-              alt="About Us"
-              width={100}
-              height={100}
-            />
+            <ISATechLogoMark className="h-24 w-auto" />
             <p className="body text-center">
               ISATech Society (ISAT U Innovators and Technopreneurs Society) is
               a student-led organization at Iloilo Science and Technology
@@ -192,11 +214,12 @@ function AboutUsEmpowermentSection() {
         </div>
         <div className="w-fill flex items-center justify-center">
           <Image
-            src="/assets/isatech-decoration-aboutus-1.png"
+            src="/assets/decorations/poststamp-sticker.png"
             alt="ISATech Society Reseach Hub Stamp"
             width={436}
             height={346}
-            className="z-1 h-full w-auto lg:h-96"
+            sizes="(min-width: 1040px) 403px, calc(100vw - 48px)"
+            className="z-1 h-auto w-full lg:h-80 lg:w-auto"
           />
         </div>
       </div>
@@ -213,11 +236,12 @@ function AboutUsInitiativesSection() {
       <div className="flex w-full max-w-6xl grid-cols-1 flex-col gap-6 py-5 lg:grid lg:grid-cols-2 lg:py-8">
         <div className="w-fill flex items-center justify-center">
           <Image
-            src="/assets/isatech-decoration-aboutus-2.png"
+            src="/assets/decorations/tagline-sticker.png"
             alt="Dream Innovate Succeed Sticker"
             width={436}
             height={303}
-            className="z-1 h-full w-auto lg:h-80"
+            sizes="(min-width: 1040px) 436px, calc(100vw - 48px)"
+            className="z-1 h-auto w-full lg:h-80 lg:w-auto"
           />
         </div>
         <div className="w-fill flex items-center justify-center">
@@ -235,142 +259,11 @@ function AboutUsInitiativesSection() {
   );
 }
 
-function AboutUsAdvisersSection() {
-  const [carouselApi, setCarouselApi] = useState<CarouselApi | null>(null);
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [totalItems, setTotalItems] = useState(0);
-
-  // Set up carousel state and auto-scroll behavior
-  useEffect(() => {
-    if (!carouselApi) return;
-
-    // Update the current index and total items when carousel changes
-    const updateCarouselState = () => {
-      setCurrentIndex(carouselApi.selectedScrollSnap());
-      setTotalItems(carouselApi.scrollSnapList().length);
-    };
-
-    updateCarouselState();
-
-    // Listen for carousel selection changes
-    carouselApi.on("select", updateCarouselState);
-
-    // Automatically scroll to the next item every 12 seconds
-    const interval = setInterval(() => {
-      carouselApi.scrollTo(
-        (carouselApi.selectedScrollSnap() + 1) %
-          carouselApi.scrollSnapList().length,
-      );
-    }, 12000);
-
-    // Cleanup listeners and interval on unmount or carouselApi change
-    return () => {
-      carouselApi.off("select", updateCarouselState);
-      clearInterval(interval);
-    };
-  }, [carouselApi]);
-
-  const scrollToIndex = (index: number) => {
-    carouselApi?.scrollTo(index);
-  };
-
-  return (
-    <section
-      className="relative flex w-full items-center justify-center px-6 py-8 md:py-6 lg:px-16 lg:py-14"
-      id="advisers"
-    >
-      {/* Animated Blobs BG */}
-      <BlobsAnimatedBackground
-        className="absolute h-full w-full"
-        blobs={adviserSectionBG}
-      />
-      <div className="flex w-full max-w-6xl flex-col gap-6 lg:flex-row">
-        <div className="flex w-full flex-col items-center justify-center gap-4 lg:max-w-2xs">
-          <h3 className="text-primary">Meet our ever supportive Advisers</h3>
-          {/* Large Screen Carousel Controls */}
-          <div className="hidden w-full gap-2 lg:flex">
-            <Button
-              size={"icon"}
-              onClick={() => scrollToIndex(currentIndex - 1)}
-            >
-              <ArrowLeftIcon />
-            </Button>
-            <Button
-              size={"icon"}
-              onClick={() => scrollToIndex(currentIndex + 1)}
-            >
-              <ArrowRightIcon />
-            </Button>
-          </div>
-        </div>
-        <div className="flex flex-col gap-4">
-          {/* Image Carousel */}
-          <Carousel
-            setApi={setCarouselApi}
-            className="w-full"
-            opts={{
-              align: "start",
-              duration: 20,
-              loop: true,
-            }}
-          >
-            <CarouselContent>
-              {advisers.map((adviser, idx) => (
-                <CarouselItem
-                  key={adviser.name + idx}
-                  className="select-none md:basis-1/2 lg:basis-1/2 xl:basis-1/3"
-                >
-                  <div className="bg-accent relative flex h-full flex-col items-center justify-center gap-2 rounded-3xl">
-                    <Image
-                      src={adviser.image}
-                      alt={adviser.name}
-                      width={300}
-                      height={400}
-                      className="h-full w-auto rounded-3xl object-cover"
-                    />
-                    <div className="from-primary to-primary/1 absolute bottom-0 h-3/12 w-full justify-start gap-2 rounded-b-3xl bg-gradient-to-t px-8 py-8 text-white">
-                      <p className="text-base font-bold">{adviser.name}</p>
-                      <p className="text-sm">{adviser.title}</p>
-                    </div>
-                  </div>
-                </CarouselItem>
-              ))}
-            </CarouselContent>
-          </Carousel>
-          {/* Mobile Carousel Controls and Indicators */}
-          <div className="flex w-full justify-center gap-2">
-            <Button
-              size={"icon"}
-              className="lg:hidden"
-              onClick={() => scrollToIndex(currentIndex - 1)}
-            >
-              <ArrowLeftIcon />
-            </Button>
-            <div className="flex w-fit items-center justify-center space-x-2">
-              {Array.from({ length: totalItems }).map((_, index) => (
-                <button
-                  key={index}
-                  onClick={() => scrollToIndex(index)}
-                  className={`h-3 w-3 rounded-full ${
-                    currentIndex === index ? "bg-primary/80" : "bg-gray-400"
-                  }`}
-                />
-              ))}
-            </div>
-            <Button
-              size={"icon"}
-              className="lg:hidden"
-              onClick={() => scrollToIndex(currentIndex + 1)}
-            >
-              <ArrowRightIcon />
-            </Button>
-          </div>
-        </div>
-      </div>
-    </section>
-  );
-}
-
+/**
+ * ################################################################################
+ * ##################################### PAGE #####################################
+ * ################################################################################
+ */
 export default function AboutUsPage() {
   return (
     <div className="flex flex-col items-center justify-center">
@@ -378,7 +271,7 @@ export default function AboutUsPage() {
       <AboutUsDescriptionSection />
       <AboutUsEmpowermentSection />
       <AboutUsInitiativesSection />
-      <AboutUsAdvisersSection />
+      <AboutUsAdvisersSection advisers={advisers} />
     </div>
   );
 }
