+++ src/app/(static)/contact/page.tsx
@@ -1,12 +1,7 @@
-"use client";
-
 import {
   ISATechDecorationLeft,
   ISATechDecorationRight,
 } from "@/components/assets/decorations";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
 import {
   LucideCog,
   LucideFacebook,
@@ -17,23 +12,34 @@ import {
   MailIcon,
 } from "lucide-react";
 import Link from "next/link";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { useTransition } from "react";
-import { sendContactEmail } from "./actions";
 import { Toaster } from "@/components/ui/sonner";
-import { toast } from "sonner";
+import { Metadata } from "next";
+import ContactUsForm from "./form";
 
-// Contact Page Configurations
+/**
+ * ################################################################################
+ * ################################### METADATA ###################################
+ * ################################################################################
+ */
+export const metadata: Metadata = {
+  title: "Contact Us | ISATech Society",
+  description:
+    "Get in touch with ISATech Society for partnerships, inquiries, or feedback.",
+  openGraph: {
+    title: "Contact Us | ISATech Society",
+    description:
+      "Get in touch with ISATech Society for partnerships, inquiries, or feedback.",
+    url: "https://isatech.club/contact",
+    siteName: "ISATech Society",
+    type: "website",
+  },
+};
+
+/**
+ * ################################################################################
+ * #################################### CONFIG ####################################
+ * ################################################################################
+ */
 const partnerList = [
   { emoji: LucideCog, text: "Industry partners for real-world projects" },
   { emoji: LucideRocket, text: "Startups for hackathon sponsorships" },
@@ -44,12 +50,16 @@ const partnerList = [
   { emoji: LucideHandshake, text: "Student clubs for cross-campus events" },
 ];
 const socialLinks = [
+  { emoji: MailIcon, text: "Email", href: "mailto:info@isatech.com" },
   { emoji: LucideFacebook, text: "Facebook", href: "https://facebook.com" },
   { emoji: LucideLinkedin, text: "LinkedIn", href: "https://linkedin.com" },
-  { emoji: MailIcon, text: "Email", href: "mailto:info@isatech.com" },
 ];
 
-// Contact Page Sections Components
+/**
+ * ################################################################################
+ * ################################## COMPONENTS ##################################
+ * ################################################################################
+ */
 function ContactUsPartnerSection() {
   return (
     <section
@@ -118,133 +128,6 @@ function ContactUsFormSection() {
   );
 }
 
-function ContactUsForm() {
-  const [loading, startTransition] = useTransition();
-
-  // Define the schema for the contact form using zod
-  const contactFormSchema = z.object({
-    // Name must be a string between 2 and 50 characters
-    name: z
-      .string()
-      .min(2, "Name must be at least 2 characters")
-      .max(50, "Name must be at most 50 characters"),
-    // Email must be a valid email address
-    email: z.email("Invalid email address"),
-    // Message must be a string between 10 and 1000 characters
-    message: z
-      .string()
-      .min(10, "Message must be at least 10 characters")
-      .max(1000, "Message must be at most 1000 characters"),
-  });
-
-  // Initialize the form using react-hook-form and zod for validation
-  const contactForm = useForm<z.infer<typeof contactFormSchema>>({
-    resolver: zodResolver(contactFormSchema), // Use zod schema for validation
-    defaultValues: {
-      name: "", // Default value for name field
-      email: "", // Default value for email field
-      message: "", // Default value for message field
-    },
-  });
-
-  async function onSubmit(values: z.infer<typeof contactFormSchema>) {
-    startTransition(async () => {
-      //form submission
-      const res = await sendContactEmail(values);
-
-      if (res.success) {
-        // Simulate successful form submission
-        contactForm.reset();
-        toast("Form submitted successfully!");
-      } else {
-        toast("Failed to submit form. Try again later.");
-      }
-    });
-  }
-
-  return (
-    <Form {...contactForm}>
-      <form
-        onSubmit={contactForm.handleSubmit(onSubmit)}
-        className="flex w-full flex-col gap-4"
-      >
-        <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
-          <FormField
-            control={contactForm.control}
-            name="name"
-            render={({ field }) => (
-              <FormItem className="flex flex-col justify-start">
-                <FormLabel htmlFor="name">Name</FormLabel>
-                <FormControl>
-                  <Input
-                    placeholder="Juan Dela Cruz"
-                    type="text"
-                    className="border-primary"
-                    autoComplete="name"
-                    id="name"
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-          <FormField
-            control={contactForm.control}
-            name="email"
-            render={({ field }) => (
-              <FormItem className="flex flex-col justify-start">
-                <FormLabel htmlFor="email">Email</FormLabel>
-                <FormControl>
-                  <Input
-                    placeholder="name@example.com"
-                    type="email"
-                    className="border-primary"
-                    autoComplete="email"
-                    id="email"
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-        </div>
-        <div className="flex flex-col gap-2">
-          <FormField
-            control={contactForm.control}
-            name="message"
-            render={({ field }) => (
-              <FormItem className="flex flex-col justify-start">
-                <FormLabel htmlFor="message">Message</FormLabel>
-                <FormControl>
-                  <Textarea
-                    placeholder="Your Message"
-                    className="border-primary field-sizing-fixed"
-                    id="message"
-                    {...field}
-                  />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-        </div>
-        <div className="flex w-full items-center justify-end">
-          <Button
-            type="submit"
-            size={"lg"}
-            variant={"default"}
-            disabled={loading}
-          >
-            {loading ? "Sending..." : "Send Message"}
-          </Button>
-        </div>
-      </form>
-    </Form>
-  );
-}
-
 function ContactUsSocialLinksSection() {
   return (
     <section className="flex w-full flex-col items-center justify-center gap-8 px-6 py-16">
@@ -276,6 +159,11 @@ function ContactUsSocialLinksSection() {
   );
 }
 
+/**
+ * ################################################################################
+ * ##################################### PAGE #####################################
+ * ################################################################################
+ */
 export default function ContactPage() {
   return (
     <main className="flex flex-col items-center justify-center">
